5491 - Move link related files
